requirements.txt
attr
cattrs
math
functools
logging
typing
pytest